refactor(external-call): extract external-call response routing
Move external-call result-disagreement routing and validation out of
TransactionConfirmationResponsesFactory into a dedicated
ExternalCallResponseRouter, matching the package's existing
checker/validator collaborator pattern. The factory returns to assembling
confirmation responses (891 -> 523 lines); the routing computation,
occurrence validation, and disagreement alarms now live in the router.

Also simplify the recorded-disagreement key match to reuse
ExternalCallKey.fromResult instead of comparing the key fields by hand.

Behavior is unchanged: the existing factory, consistency-checker, and
conformance suites (173 tests) pass before and after.
